+++ parley/src/lib.rs
@@ -35,7 +35,7 @@
 //! // Create a `RangedBuilder` or a `TreeBuilder`, which are used to construct a `Layout`.
 //! const DISPLAY_SCALE : f32 = 1.0;
 //! const TEXT : &str = "Lorem Ipsum...";
-//! let mut builder = layout_cx.ranged_builder(&mut font_cx, &TEXT, DISPLAY_SCALE);
+//! let mut builder = layout_cx.ranged_builder(&mut font_cx, &TEXT, DISPLAY_SCALE, true);
 //!
 //! // Set default styles that apply to the entire layout
 //! builder.push_default(StyleProperty::LineHeight(1.3));